requirements.txt
pandas==2.3.*
requests==2.32.*
mysql-connector-python
